pyproject.toml
[tool.poetry]
name = "weather-data-system"
version = "0.1.0"
description = ""
authors = ["VytautasPliadis <vytautas.pliadis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
sqlmodel = "^0.0.14"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
supabase = "^2.3.4"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
